Organize integration tests
This commit introduces some better organization to the integration tests
by renaming helpers.go to utils.go, moving utility methods out of the
test module (integration_test.go), and creates a new place for
constants.

This commit also removed a direct dependency on github.com/lib/pq in the
integration tests (the tests still execute successfully without this
dependency, so we can clean it up).
