+++ tests/constants.go
@@ -1,0 +1,3 @@
+package tests
+
+const clusterName = "cluster.example.com"
