--- tests/helpers.go
@@ -1,7 +1,0 @@
-package tests
-
-type Subject struct {
-	ID   string
-	Name string
-	Type string
-}
